move demo_gen out of ffi/capi and ffi/npm
