+++ ffi/npm/.gitignore
@@ -1,4 +1,3 @@
 docs
 lib
-demo
 icu_capi.wasm
