+++ ffi/npm/package.json
@@ -4,9 +4,7 @@
   "description": "ICU4X JavaScript bindings via WebAssembly",
   "main": "lib/index.mjs",
   "exports": {
-    ".": "./lib/index.mjs",
-    "./demo": "./demo/index.mjs",
-    "./demo/rendering": "./demo/rendering/rendering.mjs"
+    ".": "./lib/index.mjs"
   },
   "type": "module",
   "directories": {
