+++ tutorials/web-demo/custom/WordSegmenter.mjs
