+++ tutorials/web-demo/gen/CaseMapper.mjs
