+++ tutorials/web-demo/gen/ComposingNormalizer.mjs
