+++ tutorials/web-demo/gen/Decimal.d.ts
