+++ tutorials/web-demo/gen/Decimal.mjs
