+++ tutorials/web-demo/gen/DecimalFormatter.mjs
