+++ tutorials/web-demo/gen/DecomposingNormalizer.mjs
