+++ tutorials/web-demo/gen/ListFormatter.mjs
