+++ tutorials/web-demo/gen/WordSegmenter.mjs
