+++ tutorials/web-demo/gen/index.mjs
