+++ tutorials/web-demo/gen/rendering/rendering.mjs
